feat: Implement routing for landing page and status dashboard
Problem:
The main domain `latency.space` was displaying the status dashboard instead of the intended landing page content from `Landing.jsx`. This was because the React application served by the status container was not configured for routing and always rendered the dashboard component (`App.jsx`).

Solution:
1.  **Implement React Routing:**
    -   I refactored `status/src/App.jsx` to utilize `react-router-dom`.
    -   I extracted the status dashboard logic into a new component `status/src/pages/StatusDashboard.jsx`.
    -   I defined routes in `App.jsx`:
        -   `/`: Renders `LandingPage` component (`status/src/pages/Landing.jsx`).
        -   `/status`: Renders `StatusDashboard` component.
2.  **Update Nginx Configuration Generation:**
    -   I modified the `deploy/update-nginx.sh` script, which dynamically generates the Nginx configuration.
    -   I updated the script to proxy requests for the root domain (`latency.space`) to the root of the status container (`http://${STATUS_IP}:80/`), allowing the React router to handle the `/` path.
    -   I updated the script to proxy requests for the `status.latency.space` subdomain to the `/status` path of the status container (`http://${STATUS_IP}:80/status;/`), directing users to the status dashboard route.

Outcome:
-   Visiting `latency.space` now correctly displays the landing page.
-   Visiting `status.latency.space` now correctly displays the status dashboard.
-   The static Nginx config `config/nginx/latency.space.conf` was also updated during investigation but is superseded by the dynamic generation in `update-nginx.sh`. The dynamic script now reflects the correct routing logic.
