+++ config/nginx/latency.space.conf
@@ -238,7 +238,7 @@ server {
     
     location / {
         # Docker service resolution - using direct IP instead of DNS to avoid resolution issues
-        proxy_pass http://172.18.0.4:80;
+        proxy_pass http://172.18.0.4:80/status; # Added /status path
         
         # Standard proxy headers
         proxy_http_version 1.1;
